+++ src/Lakona.ProjectSystem/Generation/Domain/PackageCatalog.cs
@@ -23,7 +23,6 @@ internal sealed class PackageCatalog
     public string LakonaGameClient => global::GeneratedProjectPackageVersions.LakonaGameClient;
     public string LakonaGameLoadTesting => global::GeneratedProjectPackageVersions.LakonaGameLoadTesting;
     public string LakonaGameServer => global::GeneratedProjectPackageVersions.LakonaGameServer;
-    public string LakonaGameServerGenerators => global::GeneratedProjectPackageVersions.LakonaGameServerGenerators;
     public string LakonaGameCluster => global::GeneratedProjectPackageVersions.LakonaGameCluster;
     public string LakonaGameClusterRpc => global::GeneratedProjectPackageVersions.LakonaGameClusterRpc;
     public string LakonaGameClusterRpcTransportTcp => global::GeneratedProjectPackageVersions.LakonaGameClusterRpcTransportTcp;